Move makeFixtureIntoRepo to suite_test.go
It's used in more than one place already; move it to a central location.
